Optimized into one line
After thinking about it and talking to friends, I realized that mapping through the delimiters and mutating the data thrice it is an unnecessary move when a simple regexp with all three delimiters listed and their replacement value `the comma` right away is easy. Should probably add functionality to escape the split with a backslash or something like that.
TODO -> `\[delimiter]` to skip when splitting.