Initial Commit
This is used generically when getting a JSON object from a database for an application that I am working on. Often the data was initially stored in a poor state with a bunch of empty lists and arrays or random strings that need to be trimmed, so to get the data to a good starting point for whatever needs to be done with that, this recursive function makes my data mutation easier down the road.