Initial Commit
This is based on an interview question I received once. The goal is to parse a string into their number values, remove the 2 lowest values, and then get the average of the remaining numbers. I chose to use floating point values for the numbers, but that could change based on use cases. This is still a work in progress and see several areas for optimizations. Like the string concatenation is wonky, but `https://dotnetfiddle.net/` wouldn't compile with $"output => {value}";