Refactored into Class
After consideration, these methods are all very self-contained and wouldn't really be used in the application other that to find these values.
I added two methods to create instances of the class. One to set all the values to the IEnumerable and the other to also remove the lowest values and show the Console entries associated with it automatically.