Initial Commit
This was inspired by another interview question. The goal was to take a string and count the number of instances for each letter. I chose to store it as a Dictionary with the letters and the number of instances saved. A pretty fun exercise that I understood immediately but took some time to hammer out the minutia as I worked it.
I added the ability to count numbers as well. Created a method to remove special characters. Ignores spaces every time.